requirements.txt
asgiref
boto3
botocore
Django
django-storages
django-resized
django-widget-tweaks
django-social-share
django-rosetta
django-parler
django-pwa
django-embed-video
gunicorn
jmespath
pilkit
psycopg2
Pillow
python-dateutil
s3transfer
six
sqlparse
whitenoise
urllib3
